✨ feat: add a game over / restart the game
the components can now have a Game over stat which allows them to be grayed out and mark the grid slightly in order to display the game over. The keys are deactivated as well as the progress of the game. A button appears to restart the game when Game over appears in the middle of the grid

feature #19